error fix

+++ actions/dashboard/billboard/billboardsActions.tsx
@@ -1,6 +1,7 @@
 "use server";
 
 import prisma from "@/lib/db";
+import { getCurrentUser } from "@/lib/getCurrentUser";
 
 interface BillboardProps {
   imageUrl: string;
@@ -52,6 +53,16 @@ export async function updateBillboard(billboardId:string, data: BillboardProps) 
 
 
 export async function deleteBillboard(id: string) { 
+
+  const currentUser = await getCurrentUser();
+
+  if(currentUser?.role !== "SUPER_ADMIN") {
+    return {
+      status: 400,
+      message: "You don't have permission to delete this billboard.",
+    };
+  }
+
   try {
     await prisma.billboard.delete({ 
       where: {
